positions_backend_micro/requirements.txt
aws-cdk-lib==2.103.1
constructs>=10.0.0,<11.0.0

aws-cdk.aws-apigatewayv2-alpha==2.103.1a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.103.1a0